+++ test/subscriptions.tests.js
@@ -1939,53 +1939,6 @@ describe(
       );
     });
 
-    it('should consume to messages from a replyTo queue', (test, done) => {
-      const replyTo = uuid();
-      createBroker(
-        {
-          vhosts: {
-            '/': {
-              namespace,
-              exchanges: {
-                e1: {
-                  assert: true,
-                },
-              },
-              queues: {
-                q1: {
-                  assert: true,
-                  replyTo,
-                },
-              },
-              bindings: {
-                b1: {
-                  source: 'e1',
-                  destination: 'q1',
-                  bindingKey: 'foo.#',
-                },
-              },
-            },
-          },
-          publications: _.pick(publications, 'p1'),
-          subscriptions: _.pick(subscriptions, 's1'),
-        },
-        (err, broker) => {
-          assert.ifError(err);
-          broker.publish('p1', 'test message', `${replyTo}.foo.bar`, (err) => {
-            assert.ifError(err);
-            broker.subscribe('s1', (err, subscription) => {
-              assert.ifError(err);
-              subscription.on('message', (message, content, ackOrNack) => {
-                ackOrNack();
-                assert.strictEqual(content, 'test message');
-                done();
-              });
-            });
-          });
-        }
-      );
-    });
-
     it('should emit channel errors', (test, done) => {
       createBroker(
         {